fix: test isolation
