+++ crates/testing/e2e-tests/Cargo.toml
@@ -32,9 +32,10 @@ futures = { workspace = true, optional = true }
 escargot = { workspace = true }
 
 [dev-dependencies]
-reth = { workspace = true }
-tempfile = { workspace = true }
-serde_json = { workspace = true }
+    reth = { workspace = true }
+    tempfile = { workspace = true }
+    serde_json = { workspace = true }
+    serde_yaml = { workspace = true }
 rand_chacha = { workspace = true }
 alloy = { workspace = true }
 rand = { workspace = true }
